separate modules

+++ java/lance-namespace-hive2/src/test/java/com/lancedb/lance/namespace/hive2/TestHive2Namespace.java
@@ -78,8 +78,10 @@ public class TestHive2Namespace {
       metastore.stop();
     }
 
-    File file = new File(tmpDirBase);
-    file.delete();
+    if (tmpDirBase != null) {
+      File file = new File(tmpDirBase);
+      file.delete();
+    }
   }
 
   @AfterEach
